requirements.txt
# Web Frameworks and APIs
fastapi>=0.96.0
uvicorn>=0.22.0
websockets>=11.0.3
python-multipart>=0.0.6
jinja2>=3.1.2

# Browser Automation
playwright>=1.34.0

# Tool Library
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.11.0
pydantic>=2.0.0

# Performance optimization (optional)
aiodns>=3.0.0
charset-normalizer>=3.0.0
ujson>=5.8.0
psutil>=5.9.0

# Additional dependencies
aiohttp>=3.8.4
pytest>=7.3.1
pytest-asyncio>=0.21.0
markdown>=3.4.3 